# chronos_app/packages/api-documentation/src/yml/agent-protocol.openapi.yaml
openapi: 3.1.0
info:
    title: Chronos Agent Protocol
    version: 1.6.0
    summary: Wire-format contract for the Chronos Agent Registry and MCP Gateway.
    description: |
        Two-direction protocol between the Chronos control plane and external HTTP
        agents.

        - **Platform → Agent.** What an external HTTP agent must implement to be
          invokable via the Chronos Agent Registry. Health probe + an OpenAI-compatible
          chat completions endpoint with two Chronos extension fields.
        - **Agent → Platform.** What Chronos exposes for tool brokering. Agents call
          back into the MCP Gateway to invoke registered MCP tools without holding
          MCP-server credentials directly.

        See `docs/agent-registry.md` and `docs/mcp-gateway.md` for the operator
        guides. v1.6 supports the `bearer` and `header` outbound-auth shapes;
        OAuth2 client-credentials is a follow-up. v1.6 supports `streamable-http`
        and `sse` MCP transports; `stdio` is reserved for v1.8.
    license:
        name: Apache 2.0
        identifier: Apache-2.0
        url: https://github.com/popularowl/chronos/blob/main/LICENSE.md
    contact:
        name: Chronos
        email: saul@popularowl.com
servers:
    - url: http://localhost:8001
      description: External HTTP agent (illustrative — your service)
    - url: http://localhost:3000/api/v1
      description: Chronos control plane (default dev)
...
security: []
paths:
...
    /api/v1/mcp-gateway/{agentId}:
        post:
            tags:
                - mcp-gateway
            summary: MCP Streamable HTTP endpoint (client → server messages)
            description: |
                Agent-facing MCP server. Clients connect using any MCP client
                SDK and Streamable HTTP transport; this endpoint accepts
                JSON-RPC 2.0 messages from the client.

                Auth: `Authorization: Bearer <agent.mcpGatewayToken>` is
                required on `initialize` and on every subsequent message.
                The bearer is bound to the `:agentId` URL — the middleware
                cross-checks token ↔ agent and rejects mismatches with 401.

                Supported MCP methods: `initialize`, `tools/list`,
                `tools/call`. Capability advertised: `tools.listChanged: true`
                — server pushes `notifications/tools/list_changed` when the
                catalog changes.

                Detailed JSON-RPC envelope, capability negotiation, error
                codes, session-ID semantics: see the upstream
                [Model Context Protocol specification](https://modelcontextprotocol.io/specification).
                Operator-facing notes: see the agent-registry guide in repo.
            operationId: mcpGatewayClientToServer
            security:
                - mcpGatewayToken: []
            parameters:
                - in: path
                  name: agentId
                  required: true
                  schema:
                      type: string
                      format: uuid
                - in: header
                  name: Mcp-Session-Id
                  required: false
                  schema:
                      type: string
                  description: |
                      Session ID assigned by Chronos on `initialize`. Echo on
                      every subsequent message of the same session.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            description: |
                                JSON-RPC 2.0 envelope per the MCP spec —
                                `{ jsonrpc, id, method, params }`. Method
                                values: `initialize`, `tools/list`, `tools/call`.
                            type: object
            responses:
                '200':
                    description: |
                        JSON-RPC response envelope. May be a successful result
                        or a JSON-RPC error response (note: HTTP 200 with a
                        JSON-RPC error body is the protocol shape for
                        in-session failures like "tool not allowed").
                '401':
                    description: Missing / invalid Bearer, or token does not match the agent.
                '403':
                    description: Agent is BUILT_IN or disabled.
                '429':
                    description: Per-agent MCP session cap reached (`MCP_GATEWAY_MAX_SESSIONS_PER_AGENT`).
                '503':
                    description: Gateway not enabled (`ENABLE_MCP_SERVERS=false`).
        get:
            tags:
                - mcp-gateway
            summary: MCP Streamable HTTP SSE channel (server → client messages)
            description: |
                Server-to-client message channel for the MCP session. Carries
                server-pushed messages such as `notifications/tools/list_changed`.
                Opened by the client SDK after `initialize` returns a session ID.
                Refer to the [MCP specification](https://modelcontextprotocol.io/specification)
                for transport details.
            operationId: mcpGatewayServerToClient
            security:
                - mcpGatewayToken: []
            parameters:
                - in: path
                  name: agentId
                  required: true
                  schema:
                      type: string
                      format: uuid
                - in: header
                  name: Mcp-Session-Id
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: SSE stream of JSON-RPC messages from server to client.
                    content:
                        text/event-stream:
                            schema:
                                type: string
                '401':
                    description: Missing / invalid Bearer.
                '404':
                    description: Unknown session.
        delete:
            tags:
                - mcp-gateway
            summary: Explicit MCP session termination
            description: |
                Closes an active MCP session. Optional — sessions also expire
                on idle timeout (`MCP_GATEWAY_SESSION_IDLE_TIMEOUT_MS`).
            operationId: mcpGatewayCloseSession
            security:
                - mcpGatewayToken: []
            parameters:
                - in: path
                  name: agentId
                  required: true
                  schema:
                      type: string
                      format: uuid
                - in: header
                  name: Mcp-Session-Id
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Session closed.
                '401':
                    description: Missing / invalid Bearer.
                '404':
                    description: Unknown session.
...
components:
    securitySchemes:
        mcpGatewayToken:
            type: http
            scheme: bearer
            bearerFormat: hex
            description: |
                Per-agent MCP gateway token. Stored on the `Agent` row as
                `mcpGatewayToken`. Rotatable from the Agent detail UI; rotation
                is single-step (no overlap window in v1.6).
...