# chronos_app/packages/server/src/controllers/mcp-gateway/mcp.ts
import { NextFunction, Request, Response } from 'express'
import { StatusCodes } from 'http-status-codes'
import { InternalChronosError } from '../../errors/internalChronosError'
import { getRunningExpressApp } from '../../utils/getRunningExpressApp'
import { openSession, SessionCapExceededError } from '../../services/mcp-gateway-server'
import { createModuleLogger } from '../../utils/logger'
import { getErrorMessage } from '../../errors/utils'

const logger = createModuleLogger('MCPGatewayMCPController')

const requireAgent = (req: Request) => {
    const agent = req.gatewayAgent
    if (!agent) {
        throw new InternalChronosError(StatusCodes.UNAUTHORIZED, 'Gateway agent not attached to request')
    }
    return agent
}

const requireGatewayServerWiring = () => {
    const app = getRunningExpressApp()
    if (!app.mcpGateway) {
        throw new InternalChronosError(
            StatusCodes.SERVICE_UNAVAILABLE,
            'MCP gateway is not enabled. Set ENABLE_MCP_SERVERS=true to enable it.'
        )
    }
    if (!app.mcpGatewaySessionStore) {
        throw new InternalChronosError(StatusCodes.SERVICE_UNAVAILABLE, 'MCP gateway session store is not initialized.')
    }
    return {
        gateway: app.mcpGateway,
        sessionStore: app.mcpGatewaySessionStore,
        catalogChangeEmitter: app.mcpCatalogChangeEmitter
    }
}

const SESSION_HEADER = 'mcp-session-id'

/**
 * Streamable-HTTP MCP endpoint at `/api/v1/mcp-gateway/:agentId`. Handles
 * both POST (client → server JSON-RPC) and GET (server → client SSE channel).
 * Auth runs upstream via `mcpGatewayAuth`; this handler owns session lookup,
 * lazy session creation on `initialize`, and delegation to the SDK transport.
 *
 * The session ID is generated by the SDK on `initialize`; the client echoes
 * it back via the `Mcp-Session-Id` header on subsequent requests. We
 * cross-check `session.agentId === req.gatewayAgent.id` on every request as
 * defense-in-depth — `mcpGatewayAuth` already binds the bearer to the URL
 * `:agentId`, but a layered check is cheap and the security upside is
 * non-negotiable.
 */
const handle = async (req: Request, res: Response, next: NextFunction) => {
    try {
        const agent = requireAgent(req)
        const { gateway, sessionStore, catalogChangeEmitter } = requireGatewayServerWiring()

        const headerSessionId = req.headers[SESSION_HEADER]
        const sessionId =
            typeof headerSessionId === 'string' ? headerSessionId : Array.isArray(headerSessionId) ? headerSessionId[0] : undefined

        let session = sessionId ? sessionStore.get(sessionId) : undefined

        if (session && session.agentId !== agent.id) {
            // Session belongs to a different agent than the bearer authorises.
            // Should be unreachable given mcpGatewayAuth pins agent by URL,
            // but never trust a single layer.
            return res.status(StatusCodes.FORBIDDEN).json({
                jsonrpc: '2.0',
                error: { code: -32600, message: 'Session does not belong to the authenticated agent' },
                id: null
            })
        }

        if (!session) {
            // No session header (or header refers to an expired/missing session).
            // For a fresh `initialize` request this is normal — open a new session.
            // For non-initialize messages without a session header, the SDK
            // transport rejects with the appropriate JSON-RPC error.
            try {
                session = await openSession({ agent, gateway, sessionStore, catalogChangeEmitter })
            } catch (error) {
                if (error instanceof SessionCapExceededError) {
                    return res.status(StatusCodes.TOO_MANY_REQUESTS).json({
                        jsonrpc: '2.0',
                        error: {
                            code: -32000,
                            message: error.message,
                            data: { cause: 'session-cap-exceeded', limit: error.limit }
                        },
                        id: null
                    })
                }
                logger.warn(`Failed to open MCP session for agent ${agent.id}: ${getErrorMessage(error)}`)
                throw error
            }
        }

        await session.transport.handleRequest(req, res, req.body)
    } catch (error) {
        next(error)
    }
}

export default {
    handle
}